(deleted unnecessary files and modified routes)
