--- templates/max.html
@@ -1,28 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>Eulermation</title>
-  <link rel="stylesheet" href="/app/static/css/index.css">
-</head>
-<body>
-    <style>
-        body {
-          background-color: #E8FFF1 !important; /* Match your desired color */
-          margin: 0; /* Remove default margins */
-        }
-    </style>
-  <div class="pop-up" id="animation-full">
-    <div class="column2">
-        <div class="animation-box">
-            <p>Euler Circuit Animation</p>
-        </div>
-        <div class="download">
-            <button class="download-btn">⬇ Download</button>
-        </div>
-    </div>
-  </div> 
-  <script src="/app/static/js/popup.js"></script>
-</body>
-</html>
