--- templates/tailwind.config.js
@@ -1,64 +1,0 @@
-// /** @type {import('tailwindcss').Config} */
-// module.exports = {
-//     darkMode: ["class"],
-//     content: [
-//       "./pages/**/*.{js,ts,jsx,tsx,mdx}",
-//       "./components/**/*.{js,ts,jsx,tsx,mdx}",
-//       "./app/**/*.{js,ts,jsx,tsx,mdx}",
-//       "*.{js,ts,jsx,tsx,mdx}",
-//       "app/**/*.{ts,tsx}",
-//       "components/**/*.{ts,tsx}",
-//     ],
-//     theme: {
-//       extend: {
-//         colors: {
-//           border: "hsl(var(--border))",
-//           input: "hsl(var(--input))",
-//           ring: "hsl(var(--ring))",
-//           background: "hsl(var(--background))",
-//           foreground: "hsl(var(--foreground))",
-//           primary: {
-//             DEFAULT: "hsl(var(--primary))",
-//             foreground: "hsl(var(--primary-foreground))",
-//           },
-//           secondary: {
-//             DEFAULT: "hsl(var(--secondary))",
-//             foreground: "hsl(var(--secondary-foreground))",
-//           },
-//           destructive: {
-//             DEFAULT: "hsl(var(--destructive))",
-//             foreground: "hsl(var(--destructive-foreground))",
-//           },
-//           muted: {
-//             DEFAULT: "hsl(var(--muted))",
-//             foreground: "hsl(var(--muted-foreground))",
-//           },
-//           accent: {
-//             DEFAULT: "hsl(var(--accent))",
-//             foreground: "hsl(var(--accent-foreground))",
-//           },
-//           popover: {
-//             DEFAULT: "hsl(var(--popover))",
-//             foreground: "hsl(var(--popover-foreground))",
-//           },
-//           card: {
-//             DEFAULT: "hsl(var(--card))",
-//             foreground: "hsl(var(--card-foreground))",
-//           },
-//           lemon: "#E1DC53",
-//           passata: "#F16838",
-//           "fresh-leaf": "#339551",
-//           minty: "#BEEOC7",
-//           cream: "#F6F4DC",
-//           basil: "#266533",
-//         },
-//         borderRadius: {
-//           lg: "var(--radius)",
-//           md: "calc(var(--radius) - 2px)",
-//           sm: "calc(var(--radius) - 4px)",
-//         },
-//       },
-//     },
-//     plugins: [require("tailwindcss-animate")],
-//   }
-
